+++ ocdsextensionregistry/cli/commands/generate_pot_files.py
@@ -1,32 +1,20 @@
 import logging
 import subprocess
-from contextlib import closing, contextmanager
+from contextlib import closing
 from glob import glob
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
-import sphinx
 from babel.messages.catalog import Catalog
 from babel.messages.extract import extract, pathmatch
 from babel.messages.pofile import write_po
-from docutils.parsers.rst import directives
-from ocds_babel.directives import NullDirective
 from ocds_babel.extract import extract_codelist, extract_extension_metadata, extract_schema
-from recommonmark.transform import AutoStructify
 from sphinx.application import Sphinx
-from sphinx.util.docutils import docutils_namespace
+from sphinx.util.docutils import docutils_namespace, patch_docutils
 from sphinx.util.osutil import cd
 
 from ocdsextensionregistry import EXTENSION_VERSIONS_DATA, EXTENSIONS_DATA
 from ocdsextensionregistry.cli.commands.base import BaseCommand
-
-# patch_docutils is added in Sphinx 1.6. Copied from ocds-babel's translate_markdown.py.
-if sphinx.version_info >= (1, 6):
-    from sphinx.util.docutils import patch_docutils
-else:
-    @contextmanager
-    def patch_docutils(confdir=None):
-        yield
 
 logger = logging.getLogger('ocdsextensionregistry')
 
@@ -61,12 +49,8 @@ class Command(BaseCommand):
         #
         # * bin/sphinx-build calls main() in sphinx.cmd.build, which calls build_main(), which calls Sphinx(…).build(…)
 
-        if sphinx.version_info >= (1, 6):
-            # https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-suppress_warnings
-            warning_type = 'image.not_readable'
-        else:
-            # https://sphinx.readthedocs.io/en/1.5/config.html
-            warning_type = 'image.nonlocal_uri'
+        # https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-suppress_warnings
+        warning_type = 'image.not_readable'
 
         kwargs = {
             # sphinx-build -E …
@@ -82,10 +66,6 @@ class Command(BaseCommand):
         if not self.args.verbose:
             # sphinx-build -q …
             kwargs['status'] = None
-
-        # Silence warnings about unregistered directives.
-        for name in ('csv-table-no-translate', 'extensiontable'):
-            directives.register_directive(name, NullDirective)
 
         # For pybabel, the code path is:
         #
@@ -173,7 +153,7 @@ class Command(BaseCommand):
                 # This section is equivalent to running:
                 #
                 # echo -e '.. toctree::\n   :hidden:\n\n   README' > index.rst
-                # sphinx-build -v -b gettext -a -E -C -D extensions=recommonmark . outdir
+                # sphinx-build -v -b gettext -a -E -C -D extensions=myst_parser . outdir
                 # msgcat outdir/*.pot
                 with TemporaryDirectory() as srcdir:
                     infos = zipfile.infolist()
@@ -194,15 +174,11 @@ class Command(BaseCommand):
 
                         # Sphinx's config.py pop()'s extensions.
                         # https://github.com/sphinx-doc/sphinx/issues/6848
-                        kwargs['confoverrides']['extensions'] = ['recommonmark']
+                        kwargs['confoverrides']['extensions'] = ['myst_parser']
 
                         with patch_docutils(), docutils_namespace():
                             # sphinx-build -b gettext $(DOCS_DIR) $(POT_DIR)
                             app = Sphinx('.', None, 'outdir', '.', 'gettext', **kwargs)
-                            # Avoid "recommonmark_config not setted, proceed default setting".
-                            app.add_config_value('recommonmark_config', {}, True)
-                            # To extract messages from `.. list-table`.
-                            app.add_transform(AutoStructify)
                             # sphinx-build -a …
                             app.build(True)